pyproject.toml
[project]
name = "agentic-harness"
version = "0.1.0"
description = "Tool checking harness with configurable rules"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
